letzter Stand

+++ Parser/src/Parser/Parser.h
@@ -11,7 +11,7 @@ class Parser
 	
 public:
 	Parser(Scanner *scanner);
-	int processFile();
+	uint16_t processFile();
 	Token *token;
 	
 private:
@@ -20,18 +20,18 @@ private:
 	void syntaxError(uint16_t expected[], int expectedCount, Token *t);
 
 	// implement grammar
-	int PROG(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int DECLS(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int DECL(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int ARRAY(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int TYPE(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int STATEMENTS(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int STATEMENT(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int EXP(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int EXP2(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int INDEX(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int OP_EXP(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
-	int OP(Node *node, Token *myCommand[], uint8_t startCount, uint8_t tokenCount);
+	uint16_t PROG(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t DECLS(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t DECL(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t ARRAY(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t TYPE(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t STATEMENTS(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t STATEMENT(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t EXP(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t EXP2(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t INDEX(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t OP_EXP(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
+	uint16_t OP(Node *node, Token *myCommand[], uint16_t startCount, uint16_t tokenCount);
 	
 	// other things
 	Scanner *scanner;
